chapter 1, program #1
making sure we have the jave sre installed and that the JAVA_HOME and all enviroment viariables are set properly. There is Documentation here -> https://www.java.com/en/download/help/path.xml on how to accomplish this if anyone needs help with that. We are using an older jre for this class, that being the v8u261 which can be downloaded and installed from here -> https://www.oracle.com/java/technologies/javase/javase-jdk8-downloads.html. I will use the commit to specify details on each program and I don't mind any kind of feedback or constructive critisism is welcome. I am new to java and to programming so welcome any chance to optimize my thinking and coding.