chapter 11 program
this is a program that allows a user to either take or not take a
survey based on how they reply. This program is to use proper exception
handling techniques to prevent from crashing or stopping mid run. It should
allow for recovery and the continuation of operations no matter what.